+++ pkg/image/definition.go
@@ -119,8 +119,9 @@ func (d DiskSize) ToMB() int64 {
 }
 
 type RawConfiguration struct {
-	DiskSize DiskSize `yaml:"diskSize"`
-	LUKSKey  string   `yaml:"luksKey"`
+	DiskSize                 DiskSize `yaml:"diskSize"`
+	LUKSKey                  string   `yaml:"luksKey"`
+	ExpandEncryptedPartition bool     `yaml:"expandEncryptedPartition"`
 }
 
 type Packages struct {